feat: implement add task feature & refactor DI structure
- Initialize `:di` Android library module and centralize Hilt modules (`RepositoryModule`, `DataSourceModule`, `DatabaseModule`, `UseCaseModule`).
- Implement `AddTaskScreen` with `AddTaskViewModel`, supporting MVI architecture for creating tasks and subtasks.
- Update `TaskListScreen` to display real data via `TaskListViewModel` and add an empty state UI.
- Refactor `TaskRepository` and `TaskLocalDataSource` to return the inserted Row ID (`Long`).
- Update `SubTaskRepository` and `SubTaskDao` to support batch insertion of subtasks.
- Change `taskId` property type from `Int` to `Long` in `SubTaskEntity` and `SubTaskModel`.
- Configure `AppNavHost` to handle navigation between Task List and Add Task routes.
- Setup Hilt entry point in `TodoApplication` and `MainActivity`.

+++ data/src/main/java/com/ab/data/models/SubTaskEntity.kt
@@ -18,7 +18,7 @@ import androidx.room.PrimaryKey
 data class SubTaskEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
-    val taskId: Int,
+    val taskId: Long,
     val title: String,
     val isCompleted: Boolean = false
 )
